rename: NatsConnectionStatus -> CloudSyncStatus
